raqib-pcn-experiments/requirements.txt
ace-tools
awkward
awkward_pandas
dgllife
dill
igraph
jupyter
matplotlib
networkx
numpy
pandas
pickle5
plotly
scikit-learn
scikit-plot
scipy==1.9.3
seaborn
tqdm
uproot
wandb
gputil
uproot3
rdkit


# torch
# torchvision

# CUDA 11.8
#pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu124

# DGL
#pip install  dgl -f https://data.dgl.ai/wheels/torch-2.4/cu124/repo.html


